+++ builder/tools.py
@@ -2,52 +2,78 @@
 """Utility tools for story building
 """
 
-import sys
 import os
 
-from .base import ActType, Act, Must, Done, Title, Description
-from .base import Person, Stage, Item, DayTime
+from .acttypes import ActType
 
 
-def build_output_strings(story, is_debug=False):
-    '''Output strings to build.
+def build_action_strings(story, is_debug=False):
+    '''Action strings to build.
+
+    Returns:
+        action strings converted from story objects.
     '''
-    output_str = []
+    act_str = []
     
     for s in story:
-        # Symbol
         if s.act_type is ActType.SYMBOL:
-            output_str.append("## {}\n".format(s.action))
-        elif s.act_type is ActType.DESC:
-            output_str.append("{}\n".format(s.action))
-        elif s.act_type is ActType.ACT:
-            output_str.append("{}\n".format(s.action))
+            act_str.append("\n## {}\n\n".format(s.action))
+        elif s.act_type is ActType.DESC or s.act_type is ActType.ACT:
+            act_str.append("{}\n".format(s.action))
         elif s.act_type is ActType.TELL:
-            output_str.append("「{}」\n".format(s.action))
+            act_str.append("「{}」\n".format(s.action))
         elif s.act_type is ActType.THINK:
-            output_str.append("{}\n".format(s.action))
+            act_str.append("{}\n".format(s.action))
         elif s.act_type is ActType.TEST and is_debug:
-            output_str.append("> TEST:{}\n".format(s.action))
-        elif s.act_type is ActType.MUST and is_debug:
-            output_str.append("> MUST:{}\n".format(s.action))
-        elif s.act_type is ActType.DONE and is_debug:
-            output_str.append("> DONE:{}\n".format(s.action))
+            act_str.append("> TEST:{}\n".format(s.action))
         else:
             pass
 
-    return output_str
+    return act_str
 
 
-def output(story, is_debug=False):
-    '''Output story to the console.
+def build_description_strings(story, is_debug=False):
+    '''Description strings to build.
+
+    Returns:
+        description strings converted from story objects.
     '''
-    strs = build_output_strings(story, is_debug)
+    desc_str = []
+
+    for s in story:
+        if s.act_type is ActType.SYMBOL:
+            if s.description:
+                desc_str.append("\n## {} -- {}\n\n".format(s.action, s.description))
+            else:
+                desc_str.append("\n## {}\n\n".format(s.action))
+        elif s.act_type is ActType.DESC or s.act_type is ActType.ACT or s.act_type is ActType.THINK:
+            desc_str.append("{}。\n".format( _description_selected(s)))
+        elif s.act_type is ActType.TELL:
+            desc_str.append("「{}」\n".format(_description_selected(s)))
+        elif s.act_type is ActType.TEST and is_debug:
+            desc_str.append("> TEST:{}\n".format(_description_selected(s)))
+
+    return desc_str
+
+
+def output(story, is_desc=False, is_debug=False):
+    '''Output story to the console.
+
+    Returns:
+        True if complete, otherwise False.
+    '''
+    strs = build_description_strings(story, is_debug) if is_desc else build_action_strings(story, is_debug)
     for p in strs:
         print(p)
 
+    return True
 
-def output_md(story, filename='story', build_dir='build', is_debug=False):
+
+def output_md(story, filename='story', build_dir='build', is_desc=False, is_debug=False):
     '''Output story as a markdown file.
+
+    Returns:
+        True if compelete, otherwise False
     '''
     EXT_MARKDOWN = 'md'
 
@@ -56,9 +82,21 @@ def output_md(story, filename='story', build_dir='build', is_debug=False):
         os.makedirs(build_dir) # create build dir
     # create file
     filefullpath = os.path.join(build_dir, "{}.{}".format(filename, EXT_MARKDOWN))
-    strs = build_output_strings(story, is_debug)
+    strs = build_description_strings(story, is_debug) if is_desc else build_action_strings(story, is_debug)
     with open(filefullpath, 'w') as f:
         for s in strs:
             f.write(s)
 
+    return True
+
+
+def _description_selected(act):
+    '''Description selector.
+
+    Returns:
+        description if the act has a description, otherwise a action.
+    '''
+    if act.description:
+        return act.description
+    return act.action