+++ tests/test_all.py
@@ -3,20 +3,27 @@
 """
 import unittest
 
-from test_base import ClassCheckTest
-from test_tools import TestOutput
-from test_example_story01 import Story01Test
-from test_example_story01 import EpisodeTest_meeting
-from test_example_story01 import EpisodeTest_confess
+import test_base
+import test_tools
+import test_testtools
+
 
 def suite():
     '''Packing all tests.
     '''
     suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(ClassCheckTest))
-    suite.addTest(unittest.makeSuite(TestOutput))
-    suite.addTest(unittest.makeSuite(Story01Test))
-    suite.addTest(unittest.makeSuite(EpisodeTest_meeting))
-    suite.addTest(unittest.makeSuite(EpisodeTest_confess))
+
+    suite.addTest(unittest.makeSuite(test_base.SubjectTest))
+    suite.addTest(unittest.makeSuite(test_base.ActTest))
+    suite.addTest(unittest.makeSuite(test_base.TitleTest))
+    suite.addTest(unittest.makeSuite(test_base.PersonTest))
+    suite.addTest(unittest.makeSuite(test_base.StageTest))
+    suite.addTest(unittest.makeSuite(test_base.ItemTest))
+    suite.addTest(unittest.makeSuite(test_base.DayTimeTest))
+
+    suite.addTest(unittest.makeSuite(test_tools.TestTools))
+
+    suite.addTest(unittest.makeSuite(test_testtools.TestCheckTools))
+
     return suite